"New build based on revison 30760"

+++ Sources/yocto_altitude.m
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * $Id: yocto_altitude.m 28744 2017-10-03 08:14:16Z seb $
+ * $Id: yocto_altitude.m 30678 2018-04-23 15:57:29Z seb $
  *
  * Implements the high-level API for Altitude functions
  *
@@ -107,7 +107,7 @@
 -(int) setCurrentValue:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"currentValue" :rest_val];
 }
 
@@ -131,7 +131,7 @@
 -(int) setQnh:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"qnh" :rest_val];
 }
 /**
