+++ Sources/yocto_buzzer.m
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * $Id: yocto_buzzer.m 28744 2017-10-03 08:14:16Z seb $
+ * $Id: yocto_buzzer.m 30678 2018-04-23 15:57:29Z seb $
  *
  * Implements the high-level API for Buzzer functions
  *
@@ -129,7 +129,7 @@
 -(int) setFrequency:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"frequency" :rest_val];
 }
 /**
