+++ Sources/yocto_motor.m
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * $Id: yocto_motor.m 28744 2017-10-03 08:14:16Z seb $
+ * $Id: yocto_motor.m 30678 2018-04-23 15:57:29Z seb $
  *
  * Implements the high-level API for Motor functions
  *
@@ -198,7 +198,7 @@
 -(int) setDrivingForce:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"drivingForce" :rest_val];
 }
 /**
@@ -246,7 +246,7 @@
 -(int) setBrakingForce:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"brakingForce" :rest_val];
 }
 /**
@@ -297,7 +297,7 @@
 -(int) setCutOffVoltage:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"cutOffVoltage" :rest_val];
 }
 /**
@@ -399,7 +399,7 @@
 -(int) setFrequency:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"frequency" :rest_val];
 }
 /**
