+++ Sources/yocto_power.m
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * $Id: yocto_power.m 28744 2017-10-03 08:14:16Z seb $
+ * $Id: yocto_power.m 30678 2018-04-23 15:57:29Z seb $
  *
  * Implements the high-level API for Power functions
  *
@@ -125,7 +125,7 @@
 -(int) setMeter:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"meter" :rest_val];
 }
 /**
