+++ Sources/yocto_powersupply.m
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * $Id: yocto_powersupply.m 28744 2017-10-03 08:14:16Z seb $
+ * $Id: yocto_powersupply.m 30678 2018-04-23 15:57:29Z seb $
  *
  * Implements the high-level API for PowerSupply functions
  *
@@ -175,7 +175,7 @@
 -(int) setVoltageSetPoint:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"voltageSetPoint" :rest_val];
 }
 /**
@@ -219,7 +219,7 @@
 -(int) setCurrentLimit:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"currentLimit" :rest_val];
 }
 /**
@@ -500,7 +500,7 @@
 -(int) setVoltageAtStartUp:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"voltageAtStartUp" :rest_val];
 }
 /**
@@ -545,7 +545,7 @@
 -(int) setCurrentAtStartUp:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"currentAtStartUp" :rest_val];
 }
 /**
