+++ Sources/yocto_refframe.m
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * $Id: yocto_refframe.m 28744 2017-10-03 08:14:16Z seb $
+ * $Id: yocto_refframe.m 30678 2018-04-23 15:57:29Z seb $
  *
  * Implements the high-level API for RefFrame functions
  *
@@ -175,7 +175,7 @@
 -(int) setBearing:(double) newval
 {
     NSString* rest_val;
-    rest_val = [NSString stringWithFormat:@"%d",(int)floor(newval * 65536.0 + 0.5)];
+    rest_val = [NSString stringWithFormat:@"%ld",(s64)floor(newval * 65536.0 + 0.5)];
     return [self _setAttr:@"bearing" :rest_val];
 }
 /**
